test(tape): split behavior contracts
