+++ test/main/session/data/transcript.test.ts
@@ -832,6 +832,34 @@ describe('SessionTranscript', () => {
       expect(sqlitePresenter.deepchatMessageTracesTable.deleteByMessageIds).not.toHaveBeenCalled()
       expect(sqlitePresenter.deepchatMessagesTable.deleteFromOrderSeq).toHaveBeenCalledWith('s1', 2)
     })
+
+    it('defers projection deletion until every tape retraction has been appended', () => {
+      const appendEvent = vi
+        .fn()
+        .mockImplementationOnce(() => undefined)
+        .mockImplementationOnce(() => {
+          throw new Error('second retraction failed')
+        })
+      const transaction = vi.fn((operation: () => unknown) => () => operation())
+      sqlitePresenter.getDatabase = vi.fn().mockReturnValue({ transaction })
+      sqlitePresenter.deepchatTapeEntriesTable = {
+        ensureBootstrapAnchor: vi.fn(),
+        appendEvent
+      }
+      sqlitePresenter.deepchatMessagesTable.getBySession.mockReturnValue([
+        createMessageRow({ id: 'm1', order_seq: 1 }),
+        createMessageRow({ id: 'm2', order_seq: 2 }),
+        createMessageRow({ id: 'm3', order_seq: 3 })
+      ])
+
+      expect(() => store.deleteFromOrderSeq('s1', 2)).toThrow('second retraction failed')
+
+      expect(transaction).toHaveBeenCalledOnce()
+      expect(appendEvent).toHaveBeenCalledTimes(2)
+      expect(sqlitePresenter.deepchatMessagesTable.deleteFromOrderSeq).not.toHaveBeenCalled()
+      expect(sqlitePresenter.deepchatSearchDocumentsTable.deleteByMessageIds).not.toHaveBeenCalled()
+      expect(sqlitePresenter.deepchatMessageTracesTable.deleteByMessageIds).not.toHaveBeenCalled()
+    })
   })
 
   describe('trace operations', () => {
